+++ parser_library/src/workspaces/macro_cache.cpp
@@ -16,6 +16,7 @@
 
 #include <array>
 
+#include "context/special_instructions.h"
 #include "file_manager.h"
 #include "lsp/lsp_context.h"
 
@@ -33,28 +34,6 @@ macro_cache::macro_cache(const file_manager& file_mngr, std::shared_ptr<file> ma
 
 std::vector<cached_opsyn_mnemo> macro_cache_key::get_opsyn_state(context::hlasm_context& ctx)
 {
-    // List of instructions that are resolved during macro definition - therefore are affected by OPSYN
-    static constexpr std::array<context::id_index, 18> cached_instr {
-        context::id_storage::well_known::COPY,
-        context::id_storage::well_known::ASPACE,
-        context::id_storage::well_known::GBLA,
-        context::id_storage::well_known::GBLB,
-        context::id_storage::well_known::GBLC,
-        context::id_storage::well_known::LCLA,
-        context::id_storage::well_known::LCLB,
-        context::id_storage::well_known::LCLC,
-        context::id_storage::well_known::SETA,
-        context::id_storage::well_known::SETB,
-        context::id_storage::well_known::SETC,
-        context::id_storage::well_known::MEND,
-        context::id_storage::well_known::MACRO,
-        context::id_storage::well_known::MEXIT,
-        context::id_storage::well_known::AIF,
-        context::id_storage::well_known::AREAD,
-        context::id_storage::well_known::ACTR,
-        context::id_storage::well_known::AGO,
-    };
-
     std::vector<cached_opsyn_mnemo> result;
 
     context::id_index last_from;
@@ -69,8 +48,8 @@ std::vector<cached_opsyn_mnemo> macro_cache_key::get_opsyn_state(context::hlasm_
             continue;
 
         // If there is an opsyn, that aliases an instruction to be CA instruction, add it to result
-        if (std::find(cached_instr.begin(), cached_instr.end(), opcode.opcode) != cached_instr.end()
-            || std::find(cached_instr.begin(), cached_instr.end(), from) != cached_instr.end())
+        if (context::instruction_resolved_during_macro_parsing(opcode.opcode)
+            || context::instruction_resolved_during_macro_parsing(from))
             result.push_back(
                 { from, opcode.opcode, std::holds_alternative<context::macro_def_ptr>(opcode.opcode_detail) });