+++ parser_library/test/workspace/macro_cache_test.cpp
@@ -337,7 +337,7 @@ TEST(macro_cache_test, overwrite_by_inline)
 
             return macro_ref.parse_macro(*this, ac, ld);
         }
-        bool has_library(std::string_view) const override { return false; }
+        bool has_library(std::string_view, resource_location*) const override { return false; }
         std::optional<std::pair<std::string, resource_location>> get_library(std::string_view) const override
         {
             return std::nullopt;