+++ parser_library/test/workspace/processor_file_test.cpp
@@ -67,7 +67,7 @@ TEST(processor_file, parse_macro)
 
             return macro_ref.parse_macro(*this, ac, ld);
         }
-        bool has_library(std::string_view) const override { return false; }
+        bool has_library(std::string_view, resource_location*) const override { return false; }
         std::optional<std::pair<std::string, resource_location>> get_library(std::string_view) const override
         {
             return std::nullopt;